Fixing output

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,7 +7,7 @@
 
 void print_number(int n)
 {
-	int temp, pow, exp;
+	long temp, pow, exp;
 
 	exp = pow = 1;
